separate out process
